boot code and other bootstraping

+++ src/core/config.c
@@ -10,7 +10,7 @@ static void config_adjust_vm_image_addr(paddr_t load_addr)
     for (size_t i = 0; i < config.vmlist_size; i++) {
         struct vm_config* vm_config = &config.vmlist[i];
         if (!vm_config->image.separately_loaded) {
-            vm_config->image.load_addr = (vm_config->image.load_addr - BAO_VAS_BASE) + load_addr;
+            vm_config->image.load_addr = (vm_config->image.load_addr - (unsigned long) BAO_VAS_BASE) + load_addr;
         }
     }
 }
